Move the MappingFactory into irbasemapping.py (prepatory to having an IRBaseMapping class).
