+++ ambassador/ambassador/ir/irhttpmapping.py
@@ -251,28 +251,3 @@ class IRHTTPMapping (IRResource):
                     return context
 
         return None
-
-
-class MappingFactory:
-    @classmethod
-    def load_all(cls, ir: 'IR', aconf: Config) -> None:
-        config_info = aconf.get_config("mappings")
-
-        if not config_info:
-            return
-
-        assert(len(config_info) > 0)    # really rank paranoia on my part...
-
-        for config in config_info.values():
-            # ir.logger.debug("creating mapping for %s" % repr(config))
-
-            mapping = IRHTTPMapping(ir, aconf, **config)
-            ir.add_mapping(aconf, mapping)
-
-    @classmethod
-    def finalize(cls, ir: 'IR', aconf: Config) -> None:
-        # OK. We've created whatever IRMappings we need. Time to create the clusters
-        # they need.
-
-        for group in ir.groups.values():
-            group.finalize(ir, aconf)
